- add splash image for android 12+ - add normal splash for android -12
