+++ flutter_native_splash.yaml
@@ -1,0 +1,6 @@
+flutter_native_splash:
+  color: "#410FA3"
+  image: "assets/images/splash.png"
+  android_12:
+    image: assets/images/android12splash.png
+    color: "#410FA3"
