+++ lib/features/on_boarding/ui/screen/on_boarding_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class OnBoardingScreen extends StatelessWidget {
+  const OnBoardingScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold();
+  }
+}
