+++ lib/my_app.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+import 'features/on_boarding/ui/screen/on_boarding_screen.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Language Learning App',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const OnBoardingScreen(),
+    );
+  }
+}
