+++ web/splash/img/dark-1x.png
